requirements.txt
streamlit
python-dotenv
fastapi
uvicorn
python-multipart
langchain
langchain-classic
langchain-openai
langchain-mcp-adapters
langchain-community
langchain-text-splitters
faiss-cpu
mcp
pypdf
tavily-python
